Detach StreamSourceProvider
