+++ Sledge.Providers/Texture/Vmt/VmtMaterialPackage.cs
@@ -18,41 +18,6 @@ namespace Sledge.Providers.Texture.Vmt
 		private MaterialTexturePackageReference _reference;
 		private IFile _file;
 
-		public class VtfTextureStreamSourceProvider : ITextureStreamSource
-		{
-			private VtfFile _vtfFile;
-			private IFile _file;
-
-			public VtfTextureStreamSourceProvider(IFile file)
-			{
-				_file = file;
-				if (file != null)
-					_vtfFile = new VtfFile(file.Open());
-			}
-			public void Dispose()
-			{
-				_vtfFile = null;
-			}
-
-			public async Task<ICollection<Bitmap>> GetImage(string item, int maxWidth, int maxHeight)
-			{
-				return await Task.Factory.StartNew(() => new List<Bitmap>(){ _vtfFile.Images.Select(x =>
-				{
-					var data = x.GetBgra32Data();
-					Bitmap bmp = new Bitmap(x.Width, x.Height, System.Drawing.Imaging.PixelFormat.Format32bppArgb);
-					BitmapData bmpData = bmp.LockBits(new Rectangle(0, 0, bmp.Width, bmp.Height), ImageLockMode.WriteOnly, PixelFormat.Format32bppArgb);
-					Marshal.Copy(data, 0, bmpData.Scan0, data.Length);
-					bmp.UnlockBits(bmpData);
-
-					return  bmp ;
-				}).Last() });
-			}
-
-			public bool HasImage(string item)
-			{
-				return _file.NameWithoutExtension.Equals(item, StringComparison.InvariantCultureIgnoreCase);
-			}
-		}
 		public VmtMaterialPackage(TexturePackageReference reference) : base(reference.File.NameWithoutExtension, "vmt")
 		{
 			_reference = reference as MaterialTexturePackageReference;
